move OutputIndex
