+++ crates/iota-genesis-builder/src/stardust/types/mod.rs
@@ -2,5 +2,6 @@
 // SPDX-License-Identifier: Apache-2.0
 
 pub mod output_header;
+pub mod output_index;
 pub mod snapshot;
 pub mod token_scheme;
